+++ spec/spec_helper.rb
@@ -8,5 +8,5 @@ SimpleCov.start do
   add_filter '/diagnostics/'
 end
 
-$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
+$LOAD_PATH.unshift File.expand_path('../lib', __dir__)
 require 'icontrol_rest'